fix bug & create function cekData

+++ Main.py
@@ -7,21 +7,18 @@ Nama Kelompok:
 - Muhammad Raffiza Azka     1910511062
 '''
 
-import os
-import time
+from os import system
+from time import localtime
 
-
+# code struktur data Queue
 class DataPenduduk:
-    nik: str
+    NIK: str
     nama: str
     jenisKelamin: chr
     alamat: str
     tglLahir: list
     usia: int
     pekerjaan: str
-
-# code struktur data Queue
-
 
 class Queue:
     def __init__(self):
@@ -39,14 +36,14 @@ class Queue:
             hasil = True
         return hasil
 
-    def add(self, nik, nama, jenisKelamin, alamat, tglLahir, usia, pekerjaan):
+    def add(self, NIK, nama, jenisKelamin, alamat, tglLahir, usia, pekerjaan):
         if(self.isEmpty() == True):
             self.last = 0
             self.first = 0
 
             self.data.append([])
             self.data[self.last] = DataPenduduk()
-            self.data[0].nik = nik
+            self.data[0].NIK = NIK
             self.data[0].nama = nama
             self.data[0].jenisKelamin = jenisKelamin
             self.data[0].alamat = alamat
@@ -58,7 +55,7 @@ class Queue:
 
             self.data.append([])
             self.data[self.last] = DataPenduduk()
-            self.data[self.last].nik = nik
+            self.data[self.last].NIK = NIK
             self.data[self.last].nama = nama
             self.data[self.last].jenisKelamin = jenisKelamin
             self.data[self.last].alamat = alamat
@@ -67,7 +64,7 @@ class Queue:
             self.data[self.last].pekerjaan = pekerjaan
 
     def getFirst(self):
-        return self.data[self.first]
+        return self.data[0]
 
     def delete(self):
         if(self.last == 0):
@@ -76,7 +73,7 @@ class Queue:
 
         else:
             for i in range(self.first+1, self.last+1):
-                self.data[i-1].nik = self.data[i].nik
+                self.data[i-1].NIK = self.data[i].NIK
                 self.data[i-1].nama = self.data[i].nama
                 self.data[i-1].jenisKelamin = self.data[i].jenisKelamin
                 self.data[i-1].alamat = self.data[i].alamat
@@ -86,18 +83,16 @@ class Queue:
             self.last = self.last - 1
 
     def printQueue(self):
-        if(self.first != 1):
+        if(self.first != -1):
             for i in range(self.first, self.last+1):
-                print(self.data[i].nik + "\t" + self.data[i].nama)
+                print(self.data[i].NIK + "\t" + self.data[i].nama)
         else:
             print("Tidak ada data")
 
 # code struktur data BST
-
-
 class Node:
-    def __init__(self, nik, nama, jenisKelamin, alamat, tglLahir, usia, pekerjaan):
-        self.nik = nik
+    def __init__(self, NIK, nama, jenisKelamin, alamat, tglLahir, usia, pekerjaan):
+        self.NIK = NIK
         self.nama = nama
         self.jenisKelamin = jenisKelamin
         self.alamat = alamat
@@ -107,13 +102,12 @@ class Node:
         self.left = None
         self.right = None
 
-
 class BinarySearchTree:
     def __init__(self):
         self.root = None
 
-    def insert(self, nik, nama, jenisKelamin, alamat, tglLahir, usia, pekerjaan):
-        newNode = Node(nik, nama, jenisKelamin, alamat,
+    def insert(self, NIK, nama, jenisKelamin, alamat, tglLahir, usia, pekerjaan):
+        newNode = Node(NIK, nama, jenisKelamin, alamat,
                        tglLahir, usia, pekerjaan)
         if (self.root == None):
             self.root = newNode
@@ -122,7 +116,7 @@ class BinarySearchTree:
         parent = None
         while(True):
             parent = current
-            if (int(nik) < int(current.nik)):
+            if (int(NIK) < int(current.NIK)):
                 current = current.left
                 if (current == None):
                     parent.left = newNode
@@ -136,7 +130,7 @@ class BinarySearchTree:
     def display(self, root):
         if (root != None):
             self.display(root.left)
-            print("NIK              :", root.nik)
+            print("NIK              :", root.NIK)
             print("Nama             :", root.nama)
             print("Kelamin          :", root.jenisKelamin)
             print("Alamat           :", root.alamat)
@@ -146,22 +140,22 @@ class BinarySearchTree:
             print("===============================")
             self.display(root.right)
 
-    def find(self, nik):
+    def find(self, NIK):
         current = self.root
         while (current != None):
-            if (int(current.nik) == int(nik)):
+            if (int(current.NIK) == int(NIK)):
                 return True
-            elif (int(current.nik) > int(nik)):
+            elif (int(current.NIK) > int(NIK)):
                 current = current.left
             else:
                 current = current.right
         return False
 
-    def printNode(self, nik):
+    def printNode(self, NIK):
         current = self.root
         while (current != None):
-            if (int(current.nik) == int(nik)):
-                print("NIK              :", current.nik)
+            if (int(current.NIK) == int(NIK)):
+                print("NIK              :", current.NIK)
                 print("Nama             :", current.nama)
                 print("Kelamin          :", current.jenisKelamin)
                 print("Alamat           :", current.alamat)
@@ -170,19 +164,19 @@ class BinarySearchTree:
                 print("Pekerjaan        :", current.pekerjaan)
                 print("===============================")
                 return 0
-            elif (int(current.nik) > int(nik)):
+            elif (int(current.NIK) > int(NIK)):
                 current = current.left
             else:
                 current = current.right
 
-    def delete(self, nik):
+    def delete(self, NIK):
         parent = self.root
         current = self.root
         isLeftChild = False
 
-        while (int(current.nik) != int(nik)):
+        while (int(current.NIK) != int(NIK)):
             parent = current
-            if (int(current.nik) > int(nik)):
+            if (int(current.NIK) > int(NIK)):
                 isLeftChild = True
                 current = current.left
             else:
@@ -246,7 +240,7 @@ class BinarySearchTree:
 
 # fungsi untuk menghitung usia dilihat dari tanggal lahir
 def hitungUsia(tglLahir):
-    waktu = time.localtime()
+    waktu = localtime()
     usia = (waktu.tm_year-1) - int(tglLahir[2])
 
     if waktu.tm_mon == int(tglLahir[1]):
@@ -257,12 +251,26 @@ def hitungUsia(tglLahir):
 
     return usia
 
+# mengecek apakah data sudah ada atau belum
+def cekData(NIK, objekQueue, objekBST):
+    for data in objekQueue.data:
+        if NIK == data.NIK:
+            exist = True
+            break
+    else:
+        exist = False
+        
+    if not exist:
+        exist = objekBST.find(NIK)
+        
+    return exist
+
 
 # Main code
 if __name__ == "__main__":
 
     # fungsi untuk clear screen
-    def clsscr(): return os.system("cls")
+    clear = lambda: system("cls")
 
     # inisialisasi objek
     queue = Queue()
@@ -272,7 +280,7 @@ if __name__ == "__main__":
     while True:
 
         # Main menu
-        clsscr()
+        clear()
         print("===========================")
         print("            Menu")
         print("===========================")
@@ -286,34 +294,40 @@ if __name__ == "__main__":
         menu = input("Pilih menu > ")
 
         if menu == "1":
-
             # kode memasukan data baru
             while True:
-                clsscr()
+                clear()
                 print("=========================")
                 print("        Buat Data")
                 print("=========================")
-                NIK = input("NIK              : ")
-                nama = input("Nama             : ")
-                jenisKelamin = input("Jenis Kelamin    : ")
-                alamat = input("Alamat           : ")
+                NIK             = input("NIK              : ")
+                nama            = input("Nama             : ")
+                jenisKelamin    = input("Jenis Kelamin    : ")
+                alamat          = input("Alamat           : ")
                 print("- Format (DD/MM/YYYY)")
-                tglLahir = input("Tanggal Lahir    : ").split("/")
-                pekerjaan = input("Pekerjaan        : ")
+                tglLahir        = input("Tanggal Lahir    : ").split("/")
+                pekerjaan       = input("Pekerjaan        : ")
 
                 # menghitung usia berdasarkan tanggal lahir
                 usia = hitungUsia(tglLahir)
 
                 tanya = input("\nApakah anda sudah yakin [y/t] > ")
-
                 if tanya.lower() == "y":
-
-                    # memasukan data yang telah diinput ke dalam queue
-                    queue.add(NIK, nama, jenisKelamin, alamat,
-                              tglLahir, usia, pekerjaan)
-                    print("\nData anda sedang diproses, mohon tunggu...")
-                    input("Tekan ENTER untuk kembali...")
-                    break
+                    exist = cekData(NIK, queue, BST)
+                    if exist:
+                        # Data tidak berhasil diinput karena double
+                        print("\nPERHATIAN: Data yang anda masukan sudah ada")
+                        input("Tekan ENTER untuk kembali...")
+                        break
+                    
+                    if not exist:
+                        # memasukan data yang telah diinput ke dalam queue
+                        queue.add(NIK, nama, jenisKelamin, alamat,
+                                tglLahir, usia, pekerjaan)
+                        print("\nData anda sedang diproses, mohon tunggu...")
+                        input("Tekan ENTER untuk kembali...")
+                        break
+                    
                 elif tanya.lower() == "t":
                     tanya = input("Apakah anda ingin mengisi ulang [y/t] > ")
 
@@ -342,7 +356,7 @@ if __name__ == "__main__":
         elif menu == "4":
             # kode menghapus data yang sudah diproses
             while True:
-                clsscr()
+                clear()
                 print("=========================")
                 print("        Hapus Data")
                 print("=========================")
@@ -354,7 +368,7 @@ if __name__ == "__main__":
                     BST.printNode(NIK)
                     tanya = input("Hapus Data? [y/t] > ")
                     if tanya.lower() == "y":
-                        clsscr()
+                        clear()
 
                         # menghapus data pada BST
                         BST.delete(NIK)
@@ -362,7 +376,7 @@ if __name__ == "__main__":
                         input("Tekan ENTER untuk kembali...")
                         break
                     elif tanya.lower() == 't':
-                        clsscr()
+                        clear()
                         print("Data tidak jadi dihapus")
                         tanya = input("Masukkan ulang NIK? [y/t] > ")
                         if tanya.lower() == 'y':
@@ -391,7 +405,7 @@ if __name__ == "__main__":
 
         elif menu == "5":
             # kode untuk melihat data yang masih dalam proses
-            clsscr()
+            clear()
             print("====================================")
             print("             Dalam Proses")
             print("====================================")
@@ -401,27 +415,30 @@ if __name__ == "__main__":
             # mencetak isi dalam antrian proses
             queue.printQueue()
 
-            # mengambil data yang ada pada antrian pertama
-            dataPertama = queue.getFirst()
-            tanya = input("\nApakah data " + dataPertama.nik +
-                          " sudah selesai dikerjakan [y/t] > ")
+            if not queue.isEmpty():    
+                # mengambil data yang ada pada antrian pertama
+                dataPertama = queue.getFirst()
+                tanya = input("\nApakah data " + dataPertama.NIK +
+                            " sudah selesai dikerjakan [y/t] > ")
 
-            # memasukan data pertama pada antrian kedalam data yang sudah selesai dikerjakan
-            if tanya.lower() == "y":
-                BST.insert(dataPertama.nik, dataPertama.nama, dataPertama.jenisKelamin,
-                           dataPertama.alamat, dataPertama.tglLahir, dataPertama.usia, dataPertama.pekerjaan)
-                print("\nData", dataPertama.nik, "telah selesai dikerjakan")
-                input("Tekan ENTER untuk kembali...")
-                queue.delete()
-            elif tanya.lower() == "t":
-                input("Tekan ENTER untuk kembali...")
+                # memasukan data pertama pada antrian kedalam data yang sudah selesai dikerjakan
+                if tanya.lower() == "y":
+                    BST.insert(dataPertama.NIK, dataPertama.nama, dataPertama.jenisKelamin,
+                            dataPertama.alamat, dataPertama.tglLahir, dataPertama.usia, dataPertama.pekerjaan)
+                    print("\nData", dataPertama.NIK, "telah selesai dikerjakan")
+                    input("Tekan ENTER untuk kembali...")
+                    queue.delete()
+                elif tanya.lower() == "t":
+                    input("Tekan ENTER untuk kembali...")
+                else:
+                    print("\nMaaf pilihan yang anda pilih tidak tersedia")
+                    input("Tekan ENTER untuk kembali...")
             else:
-                print("\nMaaf pilihan yang anda pilih tidak tersedia")
                 input("Tekan ENTER untuk kembali...")
 
         elif menu == "6":
             # kode untuk keluar dari program
-            clsscr()
+            clear()
             exit("Terima kasih...")
         else:
             # kode jika user memilih diluar jangkauan pilihan menu
